+++ Leaf_Segmentation_py3.py
@@ -7,9 +7,8 @@ Created on Thurs Oct 31 17:11:35 2019
 @author: Guillaume Theroux-Rancourt, Matt Jenkins, J. Mason Earles
 """
 
-# NOTE FROM GTR, THE LAST CONTRIBUTOR TO THIS CODE.
 # This code is intended for a non-interactive command line use.
-# New functions for reading from file written by MJ specified in comments
+# New functions added to 'Leaf_Segmentation_Functions_py3.py' by MJ
 
 # Import libraries
 from Leaf_Segmentation_Functions_py3 import *  # Custom ML microCT functions
@@ -17,43 +16,6 @@ import sys
 import os
 import gc
 from sklearn.externals import joblib
-
-# written by MJ
-def openAndReadFile(filename):
-    #opens and reads '.txt' file made by user with instructions for program...may execute full process n times
-    #initialize empty list for lines
-    list_of_lines = []
-
-    with open(filename, 'r') as f:
-        for curline in f:
-            if curline.startswith("#"):
-                pass
-            else:
-                curline = curline.replace('\n','')
-                list_of_lines.append(curline)
-            if not curline:
-                break
-    f.close()
-    return list_of_lines
-
-# written by MJ
-def define_params(list_of_lines):
-    # Extract data from command line input
-    full_script_path = list_of_lines[0]
-    sample_name = list_of_lines[1]
-    postfix_phase = list_of_lines[2]
-    Th_phase = int(list_of_lines[3])
-    postfix_grid = list_of_lines[4]
-    Th_grid = int(list_of_lines[5])
-    nb_training_slices = int(list_of_lines[6])
-    raw_slices = list_of_lines[7]
-    rescale_factor = int(list_of_lines[8])
-    threshold_rescale_factor = int(list_of_lines[9])
-    nb_estimators = int(list_of_lines[10])
-    base_folder_name = list_of_lines[11] # replaced with full_script_path
-
-    return full_script_path, sample_name, postfix_phase, Th_phase, postfix_grid, Th_grid, nb_training_slices, raw_slices, rescale_factor, threshold_rescale_factor, nb_estimators, base_folder_name
-
 
 def main():
     # Extract data from command line input
@@ -69,22 +31,21 @@ def main():
         z = z.replace('\n','')
         filenames.append(z)
 
-    # for i in range(0,len(filenames)): # optional but nice catch for incorrect filepath or filename entry
-    #     if os.path.exists(path_to_argfile_folder+filenames[i]) == False:
-    #         print("\nAt least some of the information you entered is incorrect. Try again.\n")
-    #         permission = 1
+    # optional, but nice catch for incorrect filepath or filename entry
+    for i in range(0,len(filenames)):
+        if os.path.exists(path_to_argfile_folder+filenames[i]) == False:
+            print("\nSome of the information you entered is incorrect. Try again.\n")
+            permission = 1
 
     while j < len(filenames) and permission == 0:
         print('\nWorking on scan: '+str(j+1)+' of '+str(len(filenames))+'\n')
         #read input file and define lots of stuff
         list_of_lines = openAndReadFile(path_to_argfile_folder+filenames[j])
-        # print(list_of_lines) # comment out once built
-
         # define parameters using list_of_lines
         full_script_path, sample_name, postfix_phase, Th_phase, postfix_grid, Th_grid, nb_training_slices, raw_slices, rescale_factor, threshold_rescale_factor, nb_estimators, base_folder_name = define_params(list_of_lines)
 
-        # Set directory of functions in order to import MLmicroCTfunctions
-        path_to_script = full_script_path
+        # Set directory of functions in order to import MLmicroCTfunctions (not necessary?)
+        # path_to_script = full_script_path
 
         # Get the slice numbers into a vector of integer
         ImgJ_slices = [int(x) for x in raw_slices.split(',')]
@@ -112,13 +73,8 @@ def main():
         np.random.shuffle(labelled_slices_seq)
         labelled_slices = labelled_slices[labelled_slices_seq]
         # Set the training slices to be at least one more slices than the test slices.
-        # The last part will add 1 to even length labelled slices number, and 0 to even.
-        # This meanes than odd length will have one training slice more, and even will have two more.
-        # int(len(labelled_slices) % 2 == 0))
-
         train_slices = np.arange(0, stop=nb_training_slices)
         test_slices = np.arange(len(train_slices), stop=len(labelled_slices))
-
 # TESTING
         # print(train_slices)
         # print(test_slices)
@@ -132,7 +88,6 @@ def main():
         if len(label_stack.shape) == 4:
             label_stack = label_stack[:,:,:,0]
 #
-
 #
         # Load the stacks and downsize a copy of the binary to speed up the thickness processing
         if os.path.isfile(folder_name+'/'+sample_name+'GridPhase_invert_ds.tif') == False:
@@ -149,13 +104,11 @@ def main():
             localthick_up_save(folder_name, sample_name, keep_in_memory=False)
             localthick_stack = localthick_load_and_resize(folder_name, sample_name, threshold_rescale_factor)
 #
-
         #define image subsets for training and testing
         gridphase_train_slices_subset = labelled_slices[train_slices]
         gridphase_test_slices_subset = labelled_slices[test_slices]
         label_train_slices_subset = labelled_slices_seq[train_slices]
         label_test_slices_subset = labelled_slices_seq[test_slices]
-
 # TESTING
         # print(gridphase_train_slices_subset)
         # print(gridphase_test_slices_subset)
